Type-check the clean modules in CI and fix 25 errors
mypy's declared scope (src/agent, src/spark_cli) reported 588 errors in
56 of 111 files, and nothing ran it, so the gate did not hold.

Fixed 25 errors, clearing 15 whole files: 12 stale type: ignore comments,
Any leaking from untyped boundaries (keyring, json.loads, dict.get) now
cast at the boundary, missing annotations on accumulator variables, and
three real latent bugs:

- backup._format_size declared nbytes as int but divides it in place, so
  it is a float on every pass after the first.
- gateway waited on force_deadline while guarding on force_after, so
  mypy caught a comparison against None.
- skin_engine rebound the loop variable "data" from an earlier loop;
  renamed to skin_data.

The 41 modules still carrying errors are listed in a mypy overrides
block. The other 70 are clean and now gated by a CI job, so they cannot
regress. Fix a module, delete its line from the list.

+++ src/spark_cli/backup.py
@@ -102,7 +102,7 @@ def _safe_copy_db(src: Path, dst: Path) -> bool:
 # Backup
 # ---------------------------------------------------------------------------
 
-def _format_size(nbytes: int) -> str:
+def _format_size(nbytes: float) -> str:
     """Human-readable file size."""
     for unit in ("B", "KB", "MB", "GB"):
         if nbytes < 1024:
